+++ src/clarinetsim/reputation/StrongPenalty.java
@@ -1,3 +1,3 @@
 package clarinetsim.reputation;
 
-record StrongPenalty(int value) implements Penalty {}
+record StrongPenalty(double value) implements Penalty {}
